Moves cobra commands from cmd -> app
