+++ cmd/cainjector/BUILD.bazel
@@ -8,7 +8,7 @@ go_library(
     importpath = "github.com/jetstack/cert-manager/cmd/cainjector",
     visibility = ["//visibility:private"],
     deps = [
-        "//cmd/cainjector/cmd:go_default_library",
+        "//cmd/cainjector/app:go_default_library",
         "//pkg/logs:go_default_library",
         "//pkg/util/cmd:go_default_library",
         "@io_k8s_klog//:go_default_library",
@@ -34,7 +34,7 @@ filegroup(
     name = "all-srcs",
     srcs = [
         ":package-srcs",
-        "//cmd/cainjector/cmd:all-srcs",
+        "//cmd/cainjector/app:all-srcs",
     ],
     tags = ["automanaged"],
     visibility = ["//visibility:public"],
