+++ cmd/cainjector/app/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["start.go"],
-    importpath = "github.com/jetstack/cert-manager/cmd/cainjector/cmd",
+    importpath = "github.com/jetstack/cert-manager/cmd/cainjector/app",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/api:go_default_library",
