+++ cmd/controller/app/BUILD.bazel
@@ -2,7 +2,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["controller.go"],
+    srcs = [
+        "controller.go",
+        "start.go",
+    ],
     importpath = "github.com/jetstack/cert-manager/cmd/controller/app",
     visibility = ["//visibility:public"],
     deps = [
@@ -11,6 +14,8 @@ go_library(
         "//pkg/client/clientset/versioned/scheme:go_default_library",
         "//pkg/client/informers/externalversions:go_default_library",
         "//pkg/controller:go_default_library",
+        "//pkg/controller/acmechallenges:go_default_library",
+        "//pkg/controller/acmeorders:go_default_library",
         "//pkg/controller/certificates:go_default_library",
         "//pkg/controller/clusterissuers:go_default_library",
         "//pkg/controller/expcertificates/issuing:go_default_library",
@@ -18,20 +23,30 @@ go_library(
         "//pkg/controller/expcertificates/readiness:go_default_library",
         "//pkg/controller/expcertificates/requestmanager:go_default_library",
         "//pkg/controller/expcertificates/trigger:go_default_library",
+        "//pkg/controller/ingress-shim:go_default_library",
+        "//pkg/controller/issuers:go_default_library",
         "//pkg/feature:go_default_library",
+        "//pkg/issuer/acme:go_default_library",
         "//pkg/issuer/acme/dns/util:go_default_library",
+        "//pkg/issuer/ca:go_default_library",
+        "//pkg/issuer/selfsigned:go_default_library",
+        "//pkg/issuer/vault:go_default_library",
+        "//pkg/issuer/venafi:go_default_library",
         "//pkg/logs:go_default_library",
         "//pkg/metrics:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/feature:go_default_library",
+        "@com_github_spf13_cobra//:go_default_library",
         "@io_k8s_api//core/v1:go_default_library",
         "@io_k8s_apimachinery//pkg/api/resource:go_default_library",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
+        "@io_k8s_apimachinery//pkg/util/errors:go_default_library",
         "@io_k8s_apimachinery//pkg/util/sets:go_default_library",
         "@io_k8s_client_go//informers:go_default_library",
         "@io_k8s_client_go//kubernetes:go_default_library",
         "@io_k8s_client_go//kubernetes/scheme:go_default_library",
         "@io_k8s_client_go//kubernetes/typed/core/v1:go_default_library",
+        "@io_k8s_client_go//plugin/pkg/client/auth:go_default_library",
         "@io_k8s_client_go//rest:go_default_library",
         "@io_k8s_client_go//tools/clientcmd:go_default_library",
         "@io_k8s_client_go//tools/leaderelection:go_default_library",
